+++ packages/sdk-governance/src/bounties/child-descriptors.ts
@@ -124,6 +124,7 @@ type ChildBountiesSdkPallets = PalletsTypedef<
     // }
   },
   {},
+  {},
   {}
 >
 type ChildBountiesSdkDefinition = SdkDefinition<