+++ packages/sdk-ink/src/descriptor-types.ts
@@ -148,6 +148,7 @@ export type InkSdkPallets = PalletsTypedef<
   },
   {},
   {},
+  {},
   {}
 >