+++ packages/sdk-ink/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Fix descriptor types after update to polkadot-api@1.11.0
+
 ## 0.0.1-1 2025-02-17
 
 - Update deps